+++ prototype_kl/parser/config.go
@@ -13,6 +13,8 @@ const (
 	OpComma     = ','
 	OpQuestion  = '?'
 	OpColon     = ':'
+	OpLT        = '<'
+	OpGT        = '>'
 	OpFactorial = '!'
 	OpPercent   = '%'
 	OpPower     = '^'